misc/HtmlTestSuite/results/test-0151.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="https://crbug.com/1239395" rel="help"/>
<meta content="The height of a replaced element can depend on the default natural width 300." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>

</head><body><div id="log"/>
<!-- 915977 = 299013.12 * (919 / 300) -->
<canvas data-expected-height="915977" height="919" style="width:299013.12px"/>

<script>
checkLayout('canvas');
</script>
</body></html>
